refactor : Secure AI chat route with Clerk authentication, hide widget from guest users, and migrate Gemini logic to a dedicated service layer.
