+++ server/src/routes/ai.routes.js
@@ -1,10 +1,9 @@
 import { Router } from "express";
+import { requireAuth } from "@clerk/express";
 import { handleChat } from "../controllers/ai.controller.js";
 
 const router = Router();
 
-// We don't require authentication here so guests can also use the chat
-// POST /api/v1/ai/chat
-router.post("/chat", handleChat);
+router.post("/chat", requireAuth(), handleChat);
 
 export default router;
